api-routing/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
grpcio>=1.68.0
protobuf>=4.21.0
httpx
aiohttp>=3.9.0
requests>=2.31.0
